Split PDG targeting helpers into partial class
