+++ MechJeb2/MechJeb2.csproj
@@ -77,6 +77,7 @@
     <Compile Include="PoweredDescentGuidance\PDGEngine.cs" />
     <Compile Include="PoweredDescentGuidance\PDGTerminal.cs" />
     <Compile Include="PoweredDescentGuidance\PDGPoweredDescent.cs" />
+    <Compile Include="PoweredDescentGuidance\PDGTargeting.cs" />
     <Compile Include="Maneuver\Operation.cs" />
     <Compile Include="Maneuver\OperationAdvancedTransfer.cs" />
     <Compile Include="Maneuver\OperationApoapsis.cs" />
